Correzione degli errori negli adapter
