+++ Android/app/src/main/java/com/fsc/cicerone/adapter/AdminItineraryGlobetrotterAdapter.java
@@ -66,7 +66,7 @@ public class AdminItineraryGlobetrotterAdapter extends RecyclerView.Adapter<Admi
     }//END onBindViewHolder
 
     /**
-     * Return the length of the JSON array passed into the Adapter.
+     * Return the length of the JSON array passed into the ReviewAdapter.
      *
      * @return Length of JSON array.
      */
